backend/requirements.txt
Django==5.2.10
django-debug-toolbar==6.2.0
channels[daphne]==4.3.2
django-environ==0.12.0
django-extensions==4.1
django-allauth~=65.14.0
asgiref==3.11.0
psycopg2-binary==2.9.11
channels_redis==4.3.0
celery==5.6.2
django-celery-results==2.6.0
django-celery-beat==2.8.1
django-ninja==1.5.3
tqdm==4.67.1
requests==2.32.5
pyyaml==6.0.3
django-cors-headers~=4.9.0
django-anymail[resend]==14.0
django-post-office==3.11.0
django-prometheus==2.4.1
openai==2.21.0
google-genai==1.69.0
django-storages[s3]==1.14.6
Pillow==12.1.1
